Use URL encoding to make the IDs more human-readable
